Templatize the integer parameter for multiply/divide. Normalize return values across implementations.

+++ src/emu/ioport.cpp
@@ -2033,8 +2033,8 @@ s32 ioport_manager::frame_interpolate(s32 oldval, s32 newval)
 		return newval;
 
 	// otherwise, interpolate
-	u32 nsec_since_last = (machine().time() - m_last_frame_time).as_nsec_int();
-	return oldval + (s64(newval - oldval) * s64(nsec_since_last) / s64(m_last_delta_nsec));
+	s64 nsec_since_last = (machine().time() - m_last_frame_time).as_nsec_int();
+	return oldval + (s64(newval - oldval) * nsec_since_last / s64(m_last_delta_nsec));
 }
